Fixed bug in our mapPar implementation in Prototype.js.
It was failing to pass the indices arguments to the callback.

Why?  Because we needed to `apply` the function to a list of
arguments, and we erroneously grabbed the function Array.push, which
apparently exists (at least in SpiderMonkey?), but does _not_ do the
same thing as Array.prototype.push.